optimize redis client

+++ lib_acl_cpp/include/acl_cpp/redis/redis_result.hpp
@@ -105,10 +105,10 @@ public:
 	 * 当数据类型为 REDIS_RESULT_STRING 类型时，该函数将按内存块存放的数据存储至
 	 * 连接内存中，但需要注意防止内存溢出
 	 * @param buf {string&} 存储结果数据，内部会先调用 buf.clear()
-	 * @return {size_t} 数据的总长度
+	 * @return {int} 数据的总长度，返回值 -1 表示内部数组为空
 	 */
-	size_t argv_to_string(string& buf) const;
-	size_t argv_to_string(char* buf, size_t size) const;
+	int argv_to_string(string& buf) const;
+	int argv_to_string(char* buf, size_t size) const;
 
 	/**
 	 * 当数据类型为 REDIS_RESULT_ARRAY 类型时，该函数返回所有的数组对象
